OOP client

+++ src/client/client.cpp
@@ -44,13 +44,13 @@ SensorClient::~SensorClient() {
 }
 
 void SensorClient::send_packet(const sensor_packet& packet) {
-  size_t bytes_sent = write(m_socket_fd, &packet, sizeof(sensor_packet));
+  ssize_t bytes_sent = write(m_socket_fd, &packet, sizeof(sensor_packet));
 
   if (bytes_sent < 0) {
     throw std::runtime_error("Network write operation failed");
   }
 
-  if (bytes_sent != sizeof(sensor_packet)) {
+  if (static_cast<size_t>(bytes_sent) != sizeof(sensor_packet)) {
     throw std::runtime_error("Partial write occurred");
   }
 }
